Change numeric type since apparently msvc doesnt support it

+++ src/core/string.hpp
@@ -109,7 +109,7 @@
         void clear();
         Slice<const char> slice() const;
         template<typename... Args> static String format(const char *format_string, Args... args) {
-            ssize_t buffer_size = snprintf(NULL, 0, format_string, args...);
+            i64 buffer_size = snprintf(NULL, 0, format_string, args...);
             // String takes just buffer_size because it will automatically account for null terminator.
             // While snprintf takes buffer_size + 1 to tell it it has that amount of space to write to.
             String s = String(buffer_size);
